make HMM validator mixin
