+++ src/nemos/glm_hmm/validation.py
@@ -1,6 +1,6 @@
 """Validation classes for GLMHMM and PopulationGLMHMM models."""
 
-from dataclasses import dataclass, field
+from dataclasses import dataclass
 from typing import Any, Callable, Optional, Tuple, Union
 
 import jax
@@ -12,10 +12,10 @@ from .. import validation
 from ..base_validator import RegressorValidator
 from ..glm.params import GLMParams, GLMUserParams
 from ..glm.validation import GLMValidator
-from ..hmm.params import HMMParams
 from ..type_casting import is_pynapple_tsd
 from ..typing import DESIGN_INPUT_TYPE
 from .params import GLMHMMModelParams, GLMHMMParams, GLMHMMUserParams
+from ..hmm.validation import HMMValidatorMixin, to_hmm_params, from_hmm_params
 
 
 def has_nans_only_at_border(arr):
@@ -48,7 +48,7 @@ def to_glm_hmm_params(user_params: GLMHMMUserParams) -> GLMHMMParams:
     """
     return GLMHMMParams(
         model_params=GLMHMMModelParams(user_params[:3]),
-        hmm_params=HMMParams(*(jnp.log(p) for p in user_params[3:])),
+        hmm_params=to_hmm_params(user_params[3:]),
     )
 
 
@@ -58,11 +58,7 @@ def from_glm_hmm_params(params: GLMHMMParams) -> GLMHMMUserParams:
     Converts internal model parameters (log_scale and log probabilities)
     to user-facing parameters (scale and probabilities in regular space).
     """
-    # exponentiate and re-normalize
-    initial_prob = jnp.exp(params.hmm_params.log_initial_prob)
-    initial_prob /= initial_prob.sum()
-    transition_prob = jnp.exp(params.hmm_params.log_transition_prob)
-    transition_prob /= transition_prob.sum(axis=1, keepdims=True)
+    initial_prob, transition_prob = from_hmm_params(params.hmm_params)
     return (
         params.model_params.coef,
         params.model_params.intercept,
@@ -73,16 +69,14 @@ def from_glm_hmm_params(params: GLMHMMParams) -> GLMHMMUserParams:
 
 
 @dataclass(frozen=True, repr=False)
-class GLMHMMValidator(RegressorValidator[GLMUserParams, GLMParams]):
+class GLMHMMValidator(RegressorValidator[GLMUserParams, GLMParams], HMMValidatorMixin):
     """Validate GLM-HMM parameters and inputs."""
 
-    n_states: int = field(kw_only=True)  # keyword only and required.
     expected_param_dims: Tuple[int] = (
         2,
         1,
         1,
-        1,
-        2,
+        *HMMValidatorMixin.expected_param_dims,
     )  # (coef.ndim, intercept.ndim, scale.ndim, init_prob.ndim, transition_prob.ndim)
     to_model_params: Callable[[GLMHMMUserParams], GLMHMMParams] = to_glm_hmm_params
     from_model_params: Callable[[GLMHMMParams], GLMHMMUserParams] = from_glm_hmm_params
@@ -106,9 +100,8 @@ class GLMHMMValidator(RegressorValidator[GLMUserParams, GLMParams]):
                 "instead."
             ),
         ),
-        ("check_init_and_transition_prob_shape", None),
-        ("check_init_and_transition_prob_sum_to_1", None),
         ("check_model_params_shape", None),
+        *HMMValidatorMixin.params_validation_sequence,
         *RegressorValidator.params_validation_sequence[3:],
     )
 
@@ -148,24 +141,6 @@ class GLMHMMValidator(RegressorValidator[GLMUserParams, GLMParams]):
         err_msg = err_message_format.format(*shapes)
         return super().check_array_dimensions(params, err_msg=err_msg)
 
-    def check_init_and_transition_prob_shape(
-        self, params: GLMHMMUserParams
-    ) -> GLMHMMUserParams:
-        """Check initial and transition probabilities shape."""
-        wrapped = self.wrap_user_params(params)
-        initial_prob, transition_prob = wrapped[-2:]
-        if initial_prob.shape != (self.n_states,):
-            raise ValueError(
-                f"initial_prob must be a 1-dimensional array of shape ``({self.n_states},)``. "
-                f"Provided initial_prob is of shape ``{initial_prob.shape}`` instead."
-            )
-        if transition_prob.shape != (self.n_states, self.n_states):
-            raise ValueError(
-                f"transition_prob must be a 2-dimensional array of shape ``({self.n_states}, {self.n_states})``."
-                f"Provided transition_prob is of shape ``{transition_prob.shape}`` instead."
-            )
-        return params
-
     def check_model_params_shape(self, params: GLMHMMUserParams) -> GLMHMMUserParams:
         """Check the length of the glm parameters state axis."""
         wrapped = self.wrap_user_params(params)
@@ -184,25 +159,6 @@ class GLMHMMValidator(RegressorValidator[GLMUserParams, GLMParams]):
             raise ValueError(
                 "GLM intercept must be of shape ``(n_states,)``. "
                 f"n_states is {self.n_states} but coef has shape ``{intercept.shape}``."
-            )
-        return params
-
-    def check_init_and_transition_prob_sum_to_1(
-        self, params: GLMHMMUserParams
-    ) -> GLMHMMUserParams:
-        """Check that initial and transition probability sum to 1."""
-        wrapped = self.wrap_user_params(params)
-        initial_prob, transition_prob = wrapped[-2:]
-        if not jnp.allclose(initial_prob.sum(), 1):
-            raise ValueError(
-                f"initial_prob must sum to 1, but got sum = {initial_prob.sum()}. "
-            )
-        if not jnp.allclose(jnp.sum(transition_prob, axis=1), 1):
-            row_sums = jnp.sum(transition_prob, axis=1)
-            raise ValueError(
-                f"transition_prob matrix rows must sum to 1 over columns, but got sum = {row_sums}. "
-                f"Each row i represents the probability distribution of transitioning from state i"
-                f"and must sum to 1. "
             )
         return params
